Add Excitation and Emission Filters
Add Filter correction for microscope spectra. All correction curves are also corrected to the amount of silver mirrors used.

Filters can be applied before and after the spectrum is taken. The filters before the sample alter the excitation energy correction. The filters after the sample alter the spectral correction of the emission signal.
The microscope setup has a fairly uncommon beam splitter for taking spectra, the chroma760, where it is necessary to also correct the excitation power