Add helper function for fitting
spectra are bad outside a specific range. The light source does not generate enough emission beyond 530nm and the detectors are also generating bad values with current correction curves outside of the specified area.

Apllying this directly to the correction curves or weighting by error would be better so this function is a quick fix for usage during fitting.