Rework 2d fit routine
The 2d fit routine is typically invoked after an initial 3d fit and is more accurate on the actual wavelength which varies nonlinarly due to the grating.

Fit to the excitation and emission wavelegth is done using 10fold binning of pixels along emission and excitation respectively.

The spectrum at thie point is then fittet to get the maximum intensity using an extended range without applied binning.

Fits are done using constraints.